tek_buddy/requirements.txt
Flask==3.0.3
numpy==2.1.1
openai==1.45.0
pandas==2.2.2
psycopg2==2.9.9
python-dotenv==1.0.1
scikit_learn==1.5.2
